refactor(gui): move username to annotation slice
