+++ api-editor/gui/src/features/ui/uiSlice.ts
@@ -4,7 +4,6 @@ import { RootState } from '../../app/store';
 import { CalledAfterTarget, GroupTarget } from '../annotations/annotationSlice';
 import { AbstractPythonFilter } from '../packageData/model/filters/AbstractPythonFilter';
 import { createFilterFromString, isValidFilterToken } from '../packageData/model/filters/filterFactory';
-import {isValidUsername} from "../../common/util/validation";
 
 export interface UIState {
     showAnnotationImportDialog: boolean;
@@ -20,7 +19,6 @@ export interface UIState {
     filterString: string;
     sortingMode: SortingMode;
     batchMode: BatchMode;
-    username: string;
 }
 
 type UserAction =
@@ -136,7 +134,6 @@ export const initialState: UIState = {
     heatMapMode: HeatMapMode.None,
     sortingMode: SortingMode.Alphabetical,
     batchMode: BatchMode.None,
-    username: '',
 };
 
 // Thunks --------------------------------------------------------------------------------------------------------------
@@ -312,9 +309,6 @@ const uiSlice = createSlice({
         setBatchMode(state, action: PayloadAction<BatchMode>) {
             state.batchMode = action.payload;
         },
-        setUsername(state, action: PayloadAction<string>) {
-            state.username = action.payload;
-        },
     },
     extraReducers(builder) {
         builder.addCase(initializeUI.fulfilled, (state, action) => action.payload);
@@ -353,7 +347,6 @@ export const {
     setFilterString,
     setSortingMode,
     setBatchMode,
-    setUsername,
 } = actions;
 export const uiReducer = reducer;
 
@@ -387,5 +380,3 @@ export const selectFilter = createSelector(
 );
 export const selectSortingMode = (state: RootState): SortingMode => selectUI(state).sortingMode;
 export const selectBatchMode = (state: RootState): BatchMode => selectUI(state).batchMode;
-export const selectUsername = (state: RootState): string => selectUI(state).username;
-export const selectUsernameIsValid = (state: RootState): boolean => isValidUsername(selectUsername(state));
